Create Crime prediction in machine learning
The aim of this project to determine the type of crime according to the place and time. The dataset has been downloaded from the Kaggle ( https://www.kaggle.com/c/sf-crime ) and programming used in python. The ide or tools which I used for the project is google colaboratory and jupyter. Google colaboratory is a online tool I used this tool to train my model in the cloud.